requirements.txt
flask
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
tensorflow
keras
gunicorn
flask_fontawesome
pillow
psycopg2_binary
python_decouple
